refactor: extract AppPage and unify device routes
- Add AppPage component for shared dashboard panel/navbar chrome
- Extract MeterChart component from meter device page
- Unify device detail routes to devices/[type]/[id].vue with validation
- Delete separate ecube, inverter, meter device page directories
- Refactor all pages to use AppPage
