+++ app/pages/index.vue
@@ -1,17 +1,3 @@
 <template>
-  <UDashboardPanel>
-    <template #header>
-      <UDashboardNavbar :title="$t('home')" :ui="{ right: 'gap-3' }">
-        <template #leading>
-          <UDashboardSidebarCollapse />
-        </template>
-
-        <template #right>
-          <NotificationBell />
-        </template>
-      </UDashboardNavbar>
-    </template>
-
-    <template #body> Hello World </template>
-  </UDashboardPanel>
+  <AppPage :title="$t('home')"> Hello World </AppPage>
 </template>
