+++ app/pages/settings.vue
@@ -25,29 +25,15 @@ const links = computed(
 </script>
 
 <template>
-  <UDashboardPanel id="settings" :ui="{ body: 'lg:py-12' }">
-    <template #header>
-      <UDashboardNavbar :title="$t('settings.title')" :ui="{ right: 'gap-3' }">
-        <template #leading>
-          <UDashboardSidebarCollapse />
-        </template>
-
-        <template #right>
-          <NotificationBell />
-        </template>
-      </UDashboardNavbar>
-
+  <AppPage :title="$t('settings.title')" body-class="lg:py-12">
+    <template #toolbar>
       <UDashboardToolbar>
         <UNavigationMenu :items="links" highlight class="-mx-1 flex-1" />
       </UDashboardToolbar>
     </template>
 
-    <template #body>
-      <div
-        class="flex flex-col gap-4 sm:gap-6 lg:gap-12 w-full lg:max-w-5xl mx-auto px-4"
-      >
-        <NuxtPage />
-      </div>
-    </template>
-  </UDashboardPanel>
+    <div class="flex flex-col gap-4 sm:gap-6 lg:gap-12 w-full lg:max-w-5xl mx-auto px-4">
+      <NuxtPage />
+    </div>
+  </AppPage>
 </template>
